backend/django_api/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-spectacular
Pillow
psycopg2-binary
djoser